logical_range_filter: Separate a base of ExecuteContext into an independent file.
